pyproject.toml
[project]
name = "pdf-context-extraction"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai",
    "pydantic>=2.6",
    "pydantic-ai",
    "typer",
    "rich",
    "pandas",
    "openpyxl",
    "pillow",
    "pymupdf",
    "python-pptx",
    "tenacity",
]
